gpio: dw: make deep probe compatible
When deep probing GPIO controller, we expect struct gpio_chip::dev to be
bound to a driver.

This is currently not the case as the driver for the parent nodes
creates devices for the child nodes and never binds a driver to them.

As the child nodes have compatibles themselves too, let's match against
proper drivers.

Reported-by: David Picard <david.picard@clermont.in2p3.fr>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>